python/pyproject.toml
[project]
name = "ai-trading-agents"
version = "0.1.0"
description = "AI trading agents - Python services"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.9.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "structlog>=25.0.0",
    "pandas>=2.0.0",
    "pandas-ta>=0.3.14b",
    "redis>=5.0.0",
    "openai>=1.0.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "web3>=6.0",
    "aiohttp>=3.9",

    "opentelemetry-sdk>=1.40.0",
    "opentelemetry-exporter-otlp-proto-http>=1.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
    "ruff>=0.15.5",
    "aiosqlite>=0.20",
    "time-machine>=2.9.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
